refactor: extract import rules in a domain service
